toast service clean up
