+++ angular/src/app/shared/index.ts
@@ -2,5 +2,6 @@ export * from './shared.module';
 export * from './sample-customers';
 export * from './services/circuit/circuit.service';
 export * from './services/customer/customer.service';
+export * from './services/toast/toast.service';
 export * from './toasts/activecall.toast';
 export * from './toasts/notification.toast';
