+++ angular/src/app/shared/services/toast/toast.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ToastService } from './toast.service';
+
+describe('ToastService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: ToastService = TestBed.get(ToastService);
+    expect(service).toBeTruthy();
+  });
+});
